requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
comm==0.1.2
debugpy==1.6.5
decorator==5.1.1
python-dotenv
entrypoints==0.4
executing==1.2.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
gradio==4.12.0
gspread-dataframe==3.3.0
httplib2==0.21.0
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.9
jupyter_core==5.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.6.1
packaging==23.2
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.8.1
urllib3==1.26.14
wcwidth==0.2.6
slack_bolt
unstructured
chromadb==0.4.21
tiktoken
langsmith
langchain-cli
langchain-experimental
langchain-core
langchain-community
faiss-cpu
langchain-openai
langchainhub
langchain
langserve
playwright
beautifulsoup4
html2text
langchain-openai
apify-client 
langchain-openai 
bs4